# shared/api-schema/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "APEX API",
  "description": "The F1 data layer that lives everywhere.",
  "version": "1.0.0"
 },
 "paths": {
  "/api/season/current": {
   "get": {
    "summary": "Season Current",
    "operationId": "season_current_api_season_current_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Season"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/races": {
   "get": {
    "summary": "Races",
    "operationId": "races_api_races_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Race"
         },
         "title": "Response Races Api Races Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/races/next": {
   "get": {
    "summary": "Races Next",
    "operationId": "races_next_api_races_next_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Race"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/races/{round}": {
   "get": {
    "summary": "Race By Round",
    "operationId": "race_by_round_api_races__round__get",
    "parameters": [
     {
      "name": "round",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 30,
       "minimum": 1,
       "title": "Round"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Race"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/session/next": {
   "get": {
    "summary": "Session Next",
    "operationId": "session_next_api_session_next_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Session"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/session/{session_id}/timing": {
   "get": {
    "summary": "Session Timing",
    "operationId": "session_timing_api_session__session_id__timing_get",
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^\\d{4}-\\d{1,2}-[a-z_]+$",
       "title": "Session Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Timing"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standings/drivers": {
   "get": {
    "summary": "Standings Drivers",
    "operationId": "standings_drivers_api_standings_drivers_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DriverStanding"
         },
         "title": "Response Standings Drivers Api Standings Drivers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standings/constructors": {
   "get": {
    "summary": "Standings Constructors",
    "operationId": "standings_constructors_api_standings_constructors_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ConstructorStanding"
         },
         "title": "Response Standings Constructors Api Standings Constructors Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/drivers": {
   "get": {
    "summary": "Drivers",
    "operationId": "drivers_api_drivers_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Driver"
         },
         "title": "Response Drivers Api Drivers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/drivers/{driver_id}": {
   "get": {
    "summary": "Driver",
    "operationId": "driver_api_drivers__driver_id__get",
    "parameters": [
     {
      "name": "driver_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^[a-z0-9_]{1,40}$",
       "title": "Driver Id"
      }
     },
     {
      "name": "season",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 2100,
         "minimum": 1950
        },
        {
         "type": "null"
        }
       ],
       "description": "Defaults to the current season",
       "title": "Season"
      },
      "description": "Defaults to the current season"
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DriverDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/drivers/{driver_id}/silhouette.png": {
   "get": {
    "summary": "Driver Silhouette",
    "description": "The driver's official headshot cut-out, filled with their team colour. Square PNG, transparent background.",
    "operationId": "driver_silhouette_api_drivers__driver_id__silhouette_png_get",
    "parameters": [
     {
      "name": "driver_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^[a-z0-9_]{1,40}$",
       "title": "Driver Id"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1336,
       "minimum": 1,
       "default": 432,
       "title": "Size"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Team-colour silhouette",
      "content": {
       "image/png": {}
      }
     },
     "404": {
      "description": "Not Found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/teams": {
   "get": {
    "summary": "Teams",
    "operationId": "teams_api_teams_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Team"
         },
         "title": "Response Teams Api Teams Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/teams/{team_id}": {
   "get": {
    "summary": "Team",
    "operationId": "team_api_teams__team_id__get",
    "parameters": [
     {
      "name": "team_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^[a-z0-9_]{1,40}$",
       "title": "Team Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Team"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/status": {
   "get": {
    "summary": "Api Status",
    "operationId": "api_status_api_status_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Status"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "Race": {
    "properties": {
     "season": {
      "type": "integer",
      "title": "Season"
     },
     "round": {
      "type": "integer",
      "title": "Round"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "short_name": {
      "type": "string",
      "title": "Short Name"
     },
     "circuit_name": {
      "type": "string",
      "title": "Circuit Name"
     },
     "locality": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Locality"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "country_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country Code"
     },
     "laps_total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Laps Total"
     },
     "sessions": {
      "items": {
       "$ref": "#/components/schemas/Session"
      },
      "type": "array",
      "title": "Sessions"
     }
    },
    "type": "object",
    "required": [
     "season",
     "round",
     "name",
     "short_name",
     "circuit_name",
     "locality",
     "country",
     "country_code",
     "laps_total",
     "sessions"
    ],
    "title": "Race"
   },
   "Season": {
    "properties": {
     "year": {
      "type": "integer",
      "title": "Year"
     },
     "rounds": {
      "type": "integer",
      "title": "Rounds"
     },
     "next_round": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Round"
     }
    },
    "type": "object",
    "required": [
     "year",
     "rounds",
     "next_round"
    ],
    "title": "Season"
   },
   "Session": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "round": {
      "type": "integer",
      "title": "Round"
     },
     "type": {
      "type": "string",
      "enum": [
       "FP1",
       "FP2",
       "FP3",
       "SPRINT_QUALIFYING",
       "SPRINT",
       "QUALIFYING",
       "RACE"
      ],
      "title": "Type"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "starts_at": {
      "type": "string",
      "format": "date-time",
      "title": "Starts At"
     },
     "ends_at": {
      "type": "string",
      "format": "date-time",
      "title": "Ends At"
     },
     "state": {
      "type": "string",
      "enum": [
       "upcoming",
       "starting_soon",
       "live",
       "finished"
      ],
      "title": "State"
     }
    },
    "type": "object",
    "required": [
     "id",
     "round",
     "type",
     "label",
     "starts_at",
     "ends_at",
     "state"
    ],
    "title": "Session"
   },
...
  }
 }
}